ACM/Tests/ACM.BLTest/ProductRepositoryTest.cs
using ACM.BL;
using Microsoft.VisualStudio.TestTools.UnitTesting;

namespace ACM.BLTest
{
    [TestClass]
    public class ProductRepositoryTest
    {
        [TestMethod]
        public void RetrieveTest()
        {
            //-- arrange
            var productRepository = new ProductRepository();
            var expected = new Product(2)
            {
                ProductName = "SunFlowers",
                ProductDescription = "Assorted Size Set of 4 Bright Yellow Mini SunFlowers",
                CurrentPrice = 15.96M
            };

            //-- act
            var actual = productRepository.Retrieve(2);

            //-- assert
            Assert.AreEqual(expected.CurrentPrice, actual.CurrentPrice);
            Assert.AreEqual(expected.ProductDescription, actual.ProductDescription);
            Assert.AreEqual(expected.ProductName, actual.ProductName);
        }

        [TestMethod]
        public void SaveTestValid()
        {
            //-- arrange
            var productRepository = new ProductRepository();
            var updatedProduct = new Product(2)
            {
                CurrentPrice = 18M,
                ProductDescription = "Assorted Size Set of 4 Bright Yellow Mini Sunflowers",
                ProductName = "Sunflowers",
                HasChanges = true
            };

            //-- act
            var actual = productRepository.Save(updatedProduct);

            //-- assert
            Assert.AreEqual(true, actual);
        }

        [TestMethod]
        public void SaveTestMissingPrice()
        {
            //-- arrange
            var productRepository = new ProductRepository();
            var updatedProduct = new Product(2)
            {
                CurrentPrice = null,
                ProductDescription = "Assorted Size Set of 4 Bright Yellow Mini Sunflowers",
                ProductName = "Sunflowers",
                HasChanges = true
            };

            //-- act
            var actual = productRepository.Save(updatedProduct);

            //-- assert
            Assert.AreEqual(false, actual);
        }
    }
}
